TAMAYA-273 Fix infinite recursion in metamodel
With the lazily loaded configuration in metamodel, the logic is
susceptible to an infinite recursion if any part of the configuration
setting does a getConfiguration in the middle.  The spi delegation logic
added getConfiguration logging to all static calls as part of
5a130523ab7d3a9d33091b13f62db90ddb8fba5c in incubator-tamaya, which
caused this module to infinitely recurse and fail. This commit breaks
the chain of possible recursion.